+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # @project-version@ (@build-time@)
 
 * Fixed a bug where health information on healthbar and status text was not updated immediately (only after targeting the unit) [GH-290, Comment #5581, #5588, #5648, #5651, IM]
+* Fixed a bug in  Classic and TPC Classic where combo points from the previous target where shown on the new target when switching targets [Comment #5688].
 * Added Freezing Trap Effect from Hunter's Freezing Trap to the Auras widget as CC aura [Comment #756, #5694].
 * Updated localizations (GH-291).